API

AnyMOD.Node

AnyMOD.OthPart

AnyMOD.ParElement

AnyMOD.TechPart

AnyMOD.Tree

AnyMOD.anyModel

AnyMOD.createOptModel!

AnyMOD.plotTree

AnyMOD.reportResults

AnyMOD.setObjective!

Functions

AnyMOD.anyModel — Type

options::modOptions

lock::ReentrantLock

supTs::NamedTuple{(:lvl,:step,:sca),Tuple{Int,Tuple{Vararg{Int,N} where N},Dict{Tuple{Int,Int},Float64}}}

cInfo::Dict{Int,NamedTuple}

The core model object containing all related data and subordinate objects.

Constructor and arguments

anyModel(inDir::Union{String,Array{String,1}},outDir::String; kwargs)

inDir::Union{String,Array{String,1}}: directory of input files, also allows for provide multiple directories via an array

outDir::String: directory of output files are written to

Optional arguments, data handling

objName::String: name of the model object, will be added to the name of output files and printed during reporting, default is an empty string

csvDelim::String: specifies the delimiter used within the read-in csv files, default is a comma ,

Optional arguments, model generation

decomm::Symbol: specifies if the model should perform endogenous decommissioning, options are:

:decomm: capacities are decommissioned endogenously, once decommissioned capacities cannot be put into operation again (default)

:none: no endogenous decommissioning, operated capacities equal installed capacities

:recomm: capacities are decommissioned endogenously and can be put back into operation

interCapa::Symbol: capacity expansion can be modelled at a resolution less detailed than yearly, this options determines how capacities are distributed among the subsequent years in this case, options are:

:linear: expansion is equally distributed among years resulting in a linear increase in capacity (default)

:none: all expansion occurs in the first year

supTsLvl::Int: specifies the depth in the tree of time-steps that provides years, default is 0

shortExp::Int: intervall in years between years of capacity expansion, default is 10

redStep::Float64: scales down energy quantities within the model, can be relevant when working with reduced time-series, default is 1.0

Optional arguments, reporting (see Data files for details)

reportLvl::Int: controls the frequency of writing updates to the console, default is 2

errCheckLvl::Int: controls the frequency of checking for errors, default is 2

errWrtLvl::Int: controls the frequency of writing an error report to a csv file, default is1

Optional arguments, numerical issues (see Performance and stability for details)

coefRng::NamedTuple: specifies the maximum range of coefficients in the matrix and right-hand side of the model's underlying optimization problem, default is (mat = (1e-2,1e5), rhs = (1e-2,1e2))

checkRng::Float64: if set, reports all equations whose range exceeds the specified value, default is NaN

scaFac::NamedTuple: scales different groups of variables within the model, default is (capa = 1e1, oprCapa = 1e2, dispConv = 1e3, dispSt = 1e4, dispExc = 1e3, dispTrd = 1e3, costDisp = 1e1, costCapa = 1e2, obj = 1e0)

bound::NamedTuple: sets external bounds for all capacities and dispatch variables (both in GW) and for the objective value (in Mil. €), default is (capa = NaN, disp = NaN, obj = NaN)

avaMin::Float64: availabilities smaller than this value are set to zero, since in the Conversion capacity restriction and Storage capacity restriction availabilities are inversed, this avoids high coefficients, default is 0.01

emissionLoss::Bool: determines if losses from exchange and self-discharge of storage are subject to emissions, default is true

AnyMOD.createOptModel! — Function

Create constraints of the model's underlying optimization problem.

AnyMOD.setObjective! — Function

Set the objective of the model's underlying optimization problem.

setObjective!(obj_dic::Union{Dict{Symbol,Float64},Symbol},anyM::anyModel)

obj_dic is a key-word argument that specifies the respective objective. To enable multi-criteria optimization, it can also be a dictionary assigning a weight to each objective. So far, the only supported key-word is costs.

AnyMOD.reportResults — Function

reportResults(reportType::Symbol,anyM::anyModel; rtnOpt::Tuple = (:csv,))

Missing docstring.

Missing docstring for reportTimeSeries. Check Documenter's build log for details.

Missing docstring.

Missing docstring for printObject. Check Documenter's build log for details.

Missing docstring.

Missing docstring for printDuals. Check Documenter's build log for details.

Missing docstring.

Missing docstring for printIIS. Check Documenter's build log for details.

AnyMOD.plotTree — Function

drawNodeTree(Tree_df::DataFrame, options::modOptions; args...)

Plots a tree for all nodes provided by the set data frame and copies it to the output directory defined within options.

Options and default values:

trans = 4.5

Controls fading of color going further down the tree.

wide = fill(1.0,maximum(Tree_df[!,:lvl]))

Ratio of distances between nodes that have and do not have the same parent (separate on each level).

labelsize = 7

Size of labels in graph.

ratio = 1.0

- Aspect ratio of output graph.

Missing docstring.

Missing docstring for plotEnergyFlow. Check Documenter's build log for details.

Missing docstring.

Missing docstring for moveNode!. Check Documenter's build log for details.

Types

AnyMOD.Node — Type

bla

AnyMOD.Tree — Type

blub

AnyMOD.ParElement — Type

Type including data and additional information on parameters. Fields relate to what is provided in Parameter list and include:

name::Symbol: name of the parameter

dim::Tuple: potential dimensions of parameter data

defVal::Union{Nothing,Float64}: default value

herit::Tuple: inheritance rules for parameter, see Parameter overview for details

data::DataFrame: specified parameter data, see <a href="../data/#printObject"><code>printObject</code></a> on how to export in a readable format

AnyMOD.TechPart — Type

TechPart <: AbstractModelPart

Type used for technology parts. Parameters, variables, and constraints are assigned as dictionaries via the fields par, var, and cns, respectively. Additional fields include:

name::Tuple: name of technology as a series of nodes from the technology tree

carrier::NamedTuple: energy carriers by index assigned to technology by groups (e.g. generation, use, ...)

balLvl::NamedTuple: temporal and spatial resolution for expansion and balance of the technology

capaRestr::DataFrame: specification of capacity restrictions required for technology

actSt::Tuple: actively stored carriers altough they are not leafs by index

type::Tuple: type of technology (stock, mature, or evolving)

disAgg::Bool: if true, dispatch is modelled at expansion resolution instead of dispatch resolution

modes::Tuple: different operational modes of technology

AnyMOD.OthPart — Type

OthPart <: AbstractModelPart

Type used for 'exchange', 'trade', 'balance', 'limits', and 'objectives' model parts. Parameters, variables, and constraints are assigned as dictionaries via the fields par, var, and cns, respectively.

Missing docstring.

Missing docstring for graInfo. Check Documenter's build log for details.