CSV Tables

anyMOD.printObject — Function.

printObject(print_obj::AbstractModelElement,sets::Dict{Symbol,DataFrame},options::modOptions)

Prints the data table of a model element (parameter, variable or equation) as a csv-File to the output directory defined within options.

Options and default values:

threshold = 0.001

Variables or parameter with a value below the threshold will not be included in output table. Set to nothing, if all values should be included.

filterFunc::Union{Nothing,Function} = nothing

Defines a function to filter only certain entries of data table of the element for output.

source